Add Makefile to automate steps towards POV or PNG
The involved recipe how to create a povray file from an SCAD file
is the perfect candidate to write as Makefile.

Added a Makefile in the "test radial6/" directory that takes care
of all the tedious steps, with rules to directly output a
%.pov or %.png.

It also provides a nice template for someone new to get started.