Fix rounding errors from old OpenSCAD renderings.
This replaces rounding artifact very small values with
zero:

 * 6.12303e-017
 * -1.22461e-016
 * 1.22461e-016
 * -1.83691e-016

These were happening in the older 2014 OpenSCAD versions, while modern
OpenSCAD versions properly produce zero there.

This is a purely mechanical change without touching anything else; it
is a preparation to reduce the size of the diff when we update the CSG
with the Makefile and the latest version of povray.scad: it makes it
easier to see what changed.